@@ -8,9 +8,9 @@ import { AppComponent } from './app.component';
 import { MicroBidComponent } from './micro/micro-bid.component';
 import { AuctionServices } from './shared/auctions.services';
 import { routing } from './app.routes';
-import { CoinComponent } from './coin/coin.component';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { TokensComponent } from './tokens/tokens.component';
+import { TosComponent } from "./tos/tos.component";
 import { AuthGuard } from './shared/auth-guard.service';
 import { UserService } from './shared/user.service';
 import { PaymentService } from './shared/payment.service';
@@ -20,22 +20,21 @@ import { AccountPageComponent } from './account-page/account-page.component';
 import { ListingsComponent } from './listings/listings.component';
 import { ListingsService } from './shared/listings.service';
 import { PaymentPopupComponent } from './shared/components/payment-popup/payment-popup.component';
+import { fileSaver } from "file-saver";
 import { PaymentSuccessPopupComponent } from './shared/components/payment-success-popup/payment-success-popup.component';
 import { TicketViewComponent } from './shared/components/ticket-view/ticket-view.component';
 import { CookiePolicyComponent } from './cookie-policy/cookie-policy.component';
-import { TosComponent } from './tos/tos.component';
 import { LeaderboardComponent } from './leaderboard/leaderboard.component';
 import { AboutComponent } from './about/about.component';
 import { SupportComponent } from './support/support.component';
 import { FaqComponent } from './faq/faq.component';
 import { AffiliatesComponent } from './affiliates/affiliates.component';
 import { WinsComponent } from './wins/wins.component';
-
+ 
 @NgModule({
   declarations: [
     AppComponent,
     MicroBidComponent,
-    CoinComponent,
     TokensComponent,
     PaypalExecuteComponent,
     AccountPageComponent,
@@ -43,10 +42,10 @@ import { WinsComponent } from './wins/wins.component';
     PaymentPopupComponent,
     PaymentSuccessPopupComponent,
     CookiePolicyComponent,
-    TosComponent,
     LeaderboardComponent,
     AboutComponent,
     SupportComponent,
+    TosComponent,
     FaqComponent,
     TicketViewComponent,
     AffiliatesComponent,
@@ -60,10 +59,12 @@ import { WinsComponent } from './wins/wins.component';
     HttpClientModule,
     routing,
     BrowserAnimationsModule,
-    SharedModule
+    SharedModule,
+    fileSaver
   ],
   providers: [AuctionServices, AuthGuard, UserService, PaymentService, ListingsService],
   bootstrap: [AppComponent],
   exports: []
 })
+
 export class AppModule {}
